feat: add additional checks to functions and tape class

+++ src/functions.cpp
@@ -4,15 +4,24 @@
 
 Config readConfig(const std::string& fileName)
 {
+  constexpr size_t REQUIRED_PARAMS = 4;
   std::ifstream input(fileName);
-  std::vector< int > config;
+  if (!input)
+  {
+    throw std::runtime_error("Error: Cannot open config gile");
+  }
 
+  std::vector< int > config;
   int delay = 0;
   while (input >> delay)
   {
+    if (delay < 0)
+    {
+      throw std::logic_error("Error: Delay must be non-negative");
+    }
     config.push_back(delay);
   }
-  if (config.size() != 4)
+  if (config.size() != REQUIRED_PARAMS)
   {
     throw std::logic_error("Error: Incorrect config parameters");
   }
@@ -22,8 +31,15 @@ Config readConfig(const std::string& fileName)
 
 void clearDirectory(const std::string& path)
 {
-  for (const auto& entry: std::filesystem::directory_iterator(path))
+  if (std::filesystem::exists(path))
   {
-    std::filesystem::remove_all(entry.path());
+    for (const auto& entry: std::filesystem::directory_iterator(path))
+    {
+      std::filesystem::remove_all(entry.path());
+    }
+  }
+  else
+  {
+    throw std::logic_error("Error: Directory does not exists");
   }
 }
